move json to root
